Parse objects in state, create empty when needed
Parse the objects inside the state rather than in the test, which makes
it easier to write test. They currently all use the same schema so that
makes it simpler.
Also use the NewEmpty method of the parser so that we can create an
empty live object of the proper type when needed.