Implement TypedValue for deducedTypedValue
Create a new deducedTypedValue that doesn't have a schema, and implement
the TypedValue interface for it. This also adds tests for each of the
methods, and moved the broken top-level test to working top-level tests
since most of them have been magically fixed by this change. Some of the
tests didn't make sense (assuming lists are not atomic) and thus have
been removed.