FixTabsOrDie doesn't work well with YAML structs
Especially when the struct is in a list because:
```yaml
	- a: 1
	  b: 2
```
panics because of mixed spaces and tabs. A working alternative is to
indent with tabs, but I don't think it's very pleasant:
```yaml
	- a: 1
		b: 2
```

Update the code so that prefixed tabs from the first line are removed
from all consecutive lines, and keep it that simple.