Fix bogus conflicts comparison and reporting
A couple of things that are broken with conflicts:
- We're comparing strings with DeepEqual, which does care about the
order (and we're more considering it like a set, so it's sort-of
random).
- When we print the error, we don't distinguish between errors and
conflict errors, leading to a confusing message.
- When there is a difference we don't report specifically what it is, so
you have to match the sets yourself, which is unhelpful.

Address all of these by having a comparison function that helps us
detect what has changed, so that we can compare and report properly.