Change union to not "guess" based on difference
We're still using the difference, but only to see if something wrong was sent.

The new logic is to only trust the disicriminator, which means that the
behavior for non-discriminated union is now back to before. It also
means that we can't use that process anymore to normalize applied
unions. Applied unions are normalized by trusting the applied fields
exclusively: applying two fields doesn't work, applying a discriminator
and a different value doesn't work.