Make VersionedSet into an interface
And create a new `versionedSet` object that implements that interface. A
side effect of this change is that the versionedSet is now mostly
immutable (since there are no setters on the interface), if you want to
change a `VersionedSet`, you should probably create a new one.

This will allow us to implement two things:
1. Cached `VersionedSet`: if such an object is not replaced with a new
instance, then we don't have to reserialize it.
2. Lazy `VersionSet`: we don't have to deserialize objects until we
actually need to look at the set.

This is breaking the API.