value.Value should basically never be a pointer
Change the code so that the value.Value should never be a pointer.
Either the value is nil, in which case the underlying implementation can
indicate that, or the value is not present, in which case the
value.Value interface is nil. But we don't need a third level of
indirection.