Allow tests to specify objects in different versions
This will make it possible to parse input objects with different
schemas, and test what happens if the schema changes, especially the
topology.

A few things need to happen in order to make this possible:

1. We need to be able to parse multiple different versions. A
ParseableType can only parse one type, so we need to pass in a parser
instead. The problem with the parser was that we needed a mapping to
know what was the name of the root type. Let's use the version name as
the root type. Now we can have multiple parseable type in the test.

2. Since most test actually use the same type with multiple different
version, we add a wrapper that let's you re-use the same type with
different versions.

3. DeducedType are usually all parsed the same for all versions, so
let's also create a helper for that.

4. Now we need to know which version of the object we want to compare to
at the end. We need to specify that version in each individual test, but
that's better becasue we don't have to necessarily compare in the
version of the last operation (implicit becomes explicit).