Run non-required union tests last
Currently, we run each test twice when unions aren't required: once with
and once without them enabled. Since we run with the union enabled
first, legitimate test failures are prepended with "fails when unions
are on" which is misleading since they may not be caused by unions.

Running union tests last removes that mis-leading prefix and makes it
clear if the problem comes from union or not.