value.Equals should check types in both directions
Currently, value.Equals doesn't check that the types are different in
both direction: it checks that if lhs is of type A, then rhs must also
be of type A), but it doesn't check that if lhs is not of type A but rhs
is, then they must be different.

Make it more obvious in the test that the types must always match, and fix
the test by checking in both direction.