merge: Allow duplicate keys in lhs
This not only affects merge but also `Compare` since they use the same
algorithm/code.

Duplicates fields in a set/associative-list will now be treated as an
atomic entity within that list, and will be entirely owned by the person
who made them duplicates or who changed one of the duplicates.