support optional map keys
Optional keys of a list map (= associative lists) keys are simply left out
of the set of keys, which is different from a
key with an empty value like "" for a string and obviously also different from
a non-empty value. The comparison of values already supported that and the
comparison of list values supported lists with different number of entries.

Completely empty key field lists continue to trigger an error ("associative
list with keys has an element that omits all key fields <quoted list of
fields> (and doesn't have default values for any key fields)".

Downgrading from a version which has support for a new optional key to a
version which doesn't works as long as the optional key is not used, because
the ManagedFields don't mention the new key and field and there are no list
entries which have it set. It does not work when the new field and key are
used because the older version doesn't know that it needs to consider the new
key, as the key is not listed in the older version's OpenAPI spec.

This is considered acceptable because new fields will be alpha initially and
downgrades with an alpha feature enabled are not required to work. It is worth
calling out in release notes, though.