fieldpath: convert to blackbox testing
The tests were defined in the "fieldpath" package, which prevented using
typed.NewParser (import cycle!). Defining them in "fieldpath_test" allows
that. To avoid changing all references to the package's exported symbols,
"fieldpath_test.go" defines aliases.

typed.NewParser is more strict and finds a typo in the test schema.

Also, constructing a set with NewSet(MakePath(...)) was found to not yield
the same structs as parsing a value with a schema, so now
parser.Type(typeName).FromYAML().ToFieldSet is used instead.